pyproject.toml
[project]
name = "nanosae"
version = "0.1.0"
description = "A Simple and Fast Library for Training Sparse Autoencoders for Semantic Analysis"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.6.0",
    "hdbscan>=0.8.40",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "umap-learn>=0.5.7",
    "wandb>=0.19.11",
]
